feat: :sparkles: add code to connvert from 64 length board to 120 length board
